requirements.txt
streamlit
numpy
pandas
streamlit-extras
openai
python-dotenv
tmdb3